Clean SummerBot Files, etc.
Rewrites all of the SummerBot project into an OOP style. While it may not adhere to industry standards, it demonstrates the power of abstraction and separation of components. Three empty files were deleted.

The README.md and docs folder are updated with a checklist, and all the CAD files generated.