Clean Autonomous, Organize, Refactor
None of the code logic was touched. This only moves things around to be organized, remove all warnings, adjust methods and object references, and to prepare us for working with the competition bot. Fingers crossed, all we need to do is change the PWM constants and Chassis object, and everything will work like a charm.

Notes for future code: All autonomous routines must implement the interface and be added to the sendable chooser. An autonomous routine that schedules command should cancel when exiting. A method is provided for this, as it is the author's responsibility to exit gracefully, and not rely on Robot.java, which is just a container. Autonomous routines can access objects from Robot.java through constructor parameters, or else the multiple allocation of resources will cause a runtime error.

This commit does not get a stable tag because it is untested.